requirements.txt
requests
reusables
python-box
appdirs